Apply changes from upstream
https://github.com/qt/qtbase/commit/711f7992988ab485347ebfd6b443005c0acf61d3
https://github.com/qt/qtbase/commit/e4f71b0cb5e52b4762c4c1d681eff08376e7bc0b
https://github.com/qt/qtbase/commit/245d83bb50d3534ac14fe486a1ed9864deb9c060
https://github.com/qt/qtbase/commit/6f909a5178296855cdd53b053ced9c551a2474a6
https://github.com/qt/qtbase/commit/74c8aa76089ce0b16ddd4d560530cf6df85d397e
https://github.com/qt/qtbase/commit/6b52039866db6629747faa0aa1b57c206ed1cd6f
https://github.com/qt/qtbase/commit/421aa422af2f6b147ad076ad1736b3747abc4317
https://github.com/qt/qtbase/commit/6f2c7469f86785e6ba81fe0280210ef7275099de
https://github.com/qt/qtbase/commit/f651e6a26cb59f7b7c7671cded8a10b9f98f5c71
https://github.com/qt/qtbase/commit/0d287500be09c800fbcc8f04862d316075ced546
https://github.com/qt/qtbase/commit/3b3780748ba76ddee447d06bc6e7a5f753ee6f42
https://github.com/qt/qtbase/commit/d38f86e50b01c6dd60f5a97355031e08d6a47d18

+++ src/gui/image/qbmphandler_p.h
@@ -107,7 +107,7 @@ private:
     State state;
     BMP_FILEHDR fileHeader;
     BMP_INFOHDR infoHeader;
-    int startpos;
+    qint64 startpos;
 };
 
 QT_END_NAMESPACE
